requirements-embedding-whisper.txt
fastapi
uvicorn
aio-straico==0.0.6
sentencepiece
einops
sentence-transformers
openai-whisper
python-multipart